chore: remove duplicates file from rebase and rename files
