+++ webapp/src/views/organism/form-builder/fields/tabular-input-field-builder-wrapper.tsx
@@ -2,7 +2,7 @@ import { useState} from 'react';
 import { cn } from '@app/shadcn/util/lib';
 import { useFormState } from '@app/store/jotai/form';
 import useFormFieldsAtom from '@app/store/jotai/field-selectors';
-import TabularInputField from './TabularInputField';
+import TabularInputField from './tabular-input-field';
 import { useActiveSlideComponent } from '@app/store/jotai/active-builder-component';
 import { FieldTypes } from '@app/models/dtos/form';
 import { StandardFormFieldDto } from '@app/models/dtos/form';
